refactor(panel): 抽取共享面板组件与 PanelApi 契约
